Initial "complete" set of challenges added to the tool.
I've finished adding challenges and UI as per the current ideas I have for them. There are still features that need to be added (such as some checkboxes to turn challenges off from the HTML, remembering settings, and the like) as well as figuring out how I'm going to do the Smart Codex setting. Still, this has come a long way!