Added support for skill booster pools and setting ratings/rarities through tool.
I've added a new feature, skill booster leagues. This is still a bit of a work in progress, but think of it as an alternative to a codex. The idea is you can only use the number of copies of any given skill that you've been given. Should make things interesting! I'm planning on adding sets (that let you filter the selection of skills somewhat), but I've not gotten those implemented yet. For right now, it includes all non-PvE only skills.

I've also added dialog windows you can use to set Ratings or Rarities (for the draft feature). I've included a Rarities file for your convenience.