Skill Set editing is now implemented, which makes leagues have far more potential.
Implemented a UI for editing skill sets. This UI can save the sets to disk, of course, allowing you to build up a repository of sets. Sets themselves consist of a name, distribution of each rarity per pack, and a set of skills included in the set.

Technically, the files created are text documents that can be edited or viewed in something like Notepad, if for some reason you want to do that!