Fixed a crash error added template draft functionality!
I've fixed a crash when trying to create a booster in the Booster League Draft without sufficient rarity support. It now tries lower and higher rarities before giving up entirely and returning an incomplete pack.

I've also added a button to generate X random skills from a set disregarding pack allocation and rarity.

Finally, I've added the Template Draft functionality that has long been hinted at! It is now possible to set up a template draft and, well, do it!

In the future, I want to add the ability to save template drafts mid-go, but for now, they'll be lost when you close the program.

I've also added a few more utility functions, as well as Template Draft functions, to the Settings tab.